New assets: Added keyboard input support Now we can move the Player :D Upgrading the WarningFrame and created the ErrorFrame to control better the problems

+++ src/com/davidalmarinho/utils/Vector2.java
@@ -1,9 +1,9 @@
 package com.davidalmarinho.utils;
 
 public class Vector2 {
-    public int x, y;
+    public float x, y;
 
-    public Vector2(int x, int y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
